Radar Chart is now mostly working
The chart will open up when you open the viz.html file.  Next step I'll
work on is to add in the reading of csv file.  Then I'll work on adding
some details-on-demand and try to normalize the data for each axis
using the minimum and maximum from the dataset.